+++ server/services/custom_service.py
@@ -264,7 +264,7 @@ class CustomService(Base2Service[InstalledInfo, DownloadedInfo]):
         self.instances_info[instance].installed = None
 
         if options.purge:
-            if len(self.instances_info) < 2:
+            if not any(i.installed for i in self.instances_info.values()):
                 self.service_downloaded = False
                 await self._clear_working_dir()
                 self.models_downloaded = {}