+++ server/services/rerank_service.py
@@ -312,7 +312,7 @@ class RerankService(Base2Service[InstalledInfo, DownloadedInfo]):
         self.instances_info[instance].installed = None
 
         if options.purge:
-            if len(self.instances_info) < 2:
+            if not any(i.installed for i in self.instances_info.values()):
                 self.service_downloaded = False
                 await self.docker_service.remove_image(_const.images["cpu"].name)
                 await self.docker_service.remove_image(_const.images["gpu"].name)